Phase 2/UseCase2_Ec2/ETL_PIPELINE/requirements.txt
aiobotocore==1.4.2
aiohttp==3.8.1
aioitertools==0.8.0
aiosignal==1.2.0
asn1crypto==1.4.0
async-timeout==4.0.1
attrs==21.2.0
backcall==0.2.0
boto3==1.20.13
botocore==1.20.106
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cramjam==2.5.0
cryptography==3.4.8
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
fastparquet==0.7.1
frozenlist==1.2.0
fsspec==2021.11.0
greenlet==1.1.2
idna==3.3
ipykernel==6.5.0
ipython==7.29.0
jedi==0.18.0
jmespath==0.10.0
jupyter-client==7.0.6
jupyter-core==4.9.1
matplotlib-inline==0.1.3
multidict==5.2.0
mysql-connector-python==8.0.27
nest-asyncio==1.5.1
numpy==1.21.4
oscrypto==1.2.1
pandas==1.3.4
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.22
protobuf==3.19.1
py==1.11.0
pyarrow==6.0.1
pycparser==2.21
pycryptodomex==3.11.0
Pygments==2.10.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyodbc==4.0.32
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pywin32==302
pyzmq==22.3.0
requests==2.26.0
retry==0.9.2
s3fs==2021.11.0
s3transfer==0.5.0
six==1.16.0
slack==0.0.2
slackclient==2.9.3
snowflake-connector-python==2.7.1
SQLAlchemy==1.4.26
thrift==0.15.0
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.0
urllib3==1.26.7
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.7.2
